// YOUCOM.ReserVook/src/app/features/web/master/siteconvert/model/siteconvert.model.ts
export class SiteConvertInfo {
  companyNo: string;
  scCd: string;
  scSiteCd: string;
  scSiteNm: string;
  travelAgncCd: string;
  scPositionMan: string;
  scPositionWoman: string;
  scPositionChildA: string;
  scPositionChildB: string;
  scPositionChildC: string;
  scPositionChildD: string;
  scPositionChildE: string;
  scPositionChildF: string;
  scPositionChildOther: string;
  scPersonCalcSeg: string;
  updateCnt: number;
  programId: string;
  createClerk: string;
  createMachineNo: string;
  createMachine: string;
  createDatetime: string;
  updateClerk: string;
  updateMachineNo: string;
  updateMachine: string;
  updateDatetime: string;
  status: string;

  cdName: string;
  travelAgncName: string;
}